Update patch set 1
Patch Set 1:

I have been investigating a regression that is causing the Mms app to hang (causing a hung app toast to appear repeatedly) on my phone.  I rolled back just prior to this commit and the problem goes away.  Rolling forward to include just this commit and the problem comes back.

This is on my phone with 10,000 or so messages in 100-200 threads.  I'm not in a position to explain why this change might be causing it, but I suggest this be looked at further by someone more capable than me.

Thanks.

Patch-set: 1