Add menu item to access Cell Broadcasts from MMS app.
Add a menu item to MMS to jump to the main activity of the
Cell Broadcast app when the "config_cellBroadcastAppLinks"
config item is set to true. This allows the Cell Broadcasts
launcher icon to be removed while still allowing the user to
view the list of previously received cell broadcasts.

The menu item will not be added if the CellBroadcastReceiver
package is disabled or not installed.

Bug: 6709985
Change-Id: Iaf6ec9106bbd2875da81c929945342c5649fbe6f