Sending multiple addresses via intent will result in a single address
Bug 6728637, 6736747

We need to append the addresses one-at-a-time for them to be dealt with
by chips. Also, if you manually type in multiple numbers (vs. choosing
multiple chip suggestions), only the last number will be selected. This
is because when the user types a comma to start a new number, the chips
code asks the text editor for the last token. The mms recipient editor's
token parser looked at the position of the cursor, saw it was a comma,
and returned that position. With this change, if the cursor is sitting on
a delimiter, it will back up to find the previous token. This fix actually
fixes another odd behavior. Prior to this change, if you hit backspace on
a chip, it would turn into the number and you had to delete all the chars
of the number to delete the chip. With this change, the chip is deleted in
a single backspace. Note that this token parsing didn't change in JB, but
the interaction with Chips resulted in this bug.

Change-Id: I9de569a803ba319c5ac0576a649fcf057d94c133