Mms: On attachment size exceeded, show error message
When forwarding an MMS message which is more than
the maximum allowed size, message size is checked
by a call to checkMessageSize. Upon message size
less than 0 or greater than allowed maximum message
size, ExceedMessageSizeException is thrown. Any
error's occuring while sending MMS, mms app is
implemented to crash!

With this patch, atleast ExceedMessageSizeException will be
catched and an error message will be shown to the
user instead of allowing the mms app to crash.

Change-Id: Ib165ff7003c2708e775daa02f9a4bbe227f028e1
Author: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Xiaokang Qin <xiaokang.qin@intel.com>
Signed-off-by: Beare, Bruce J <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 38953