Delete obsolete threads when drafts updated
The delete obsolete threads action in CL will be executed
before the drafts saved in CMA, since saveDraft was called
in CMA' onStop(), but the delete action was in the CL's onStart().

It's useless to getSavingDraft as the condition to check if we should
delete the obsolete threads in mDeleteObsoleteThreadsRunnable,
since it always return false at that time. We should sync to
delete the obsolete threads in the CallBack instead.

Change-Id: Ia58dca6c8e0acb728bfa122a96bf6a4351f44a51
Signed-off-by: Roger Chen <cxr514033970@gmail.com>