Update patch set 3
Patch Set 3:

The bug is when opening a draft, you shouldn't be able to edit the recipients. How do you get it in that state? Once I selected the recipients in step 1, went back the the conversation list, and selected the draft, the recipients are no longer editable. What are the steps to get the draft's recipients editable?

Patch-set: 3
Reviewer: Tom Taylor <1003618@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Verified=0