Update patch set 3
Patch Set 3:

What strange it is! It only can repro on my Nexus S flashed AOSP(master) nightly build, but it looks fine on My Galaxy Nexus (JB JR003C).

The mRecipientEditor is null when we hide it, but it excatly showed in the top panel.

You can try it if you have a Nexus S using the AOSP master build.

Patch-set: 3