We sometime need a new thread id to update drafts
We are saving the draft in a conversation which had been
deleted all messages one by one(Now the thread id is orphaned).
The draft will be saved in a obseleted thread, so we lose them
and cann't find them in ConversationList. We need to get a new
thread id in this case.

Refer to ef3eb49cde359d36f9536dbaffe5c16f3639c1f2, it fixed this
issue while we are saving the Mms draft, but Sms.

Change-Id: I29b3c1abf373877924b1afb45027417a12a1ec68
Signed-off-by: Roger Chen <cxr514033970@gmail.com>