Implement group mms
Bug 6282172

When multiple recipients are added on a new conversation, the message is
automatically turned into an mms. This is the default, but add an mms_config.xml
option to turn that feature off. In a group mms message list, show the
sender's name on the line with the timestamp.

Change-Id: I8be4c325023bcbaf2498b8c9d8346a2a94ef63ef