SmsReceiverService NullPointerException in replaceFormFeeds
Bug 7006780
Bug 7082042

Protect against a null message. When there's a null message, put an empty
message in the database (vs. "null"). While testing, I hit a crash in
MessagingNotification.updateNotification where calling:
        NotificationInfo mostRecentNotification = sNotificationSet.first();
hit an exception. We should never be inside updateNotification without
there actually being a notification. updateNotification is called from
blockingUpdateNewMessageIndicator() which is called by several async
services, such as SmsReceiverService and TransactionService. The code
relied on a global variable. There's no need for the variable to be global.

Change-Id: I23b0a0ebed30905468832b2f0f80f2f0016768f0