On receiving a Group MMS the List view displays as two messages received
Bug 7081327

When an mms is sent, the first thing that is received is an mms
notification message. This message only contains the address of
the sender and not all the other addresses the message was additionally
sent to. The messaging app stores this message in a new thread
(if there isn't an existing one) of the sender's phone number.
Then if the mms auto download setting is turned on (the default),
the app will use the information in this notification message to
download the content of the message. The app will then receive an
mms retrieve confirmation message. This message contains the
full addresses of all those in the group conversation. The app then
creates (if there isn't an existing one) a thread made up of all
the addresses. Then it deletes the original mms notification message.
At this point, you see two two threads for the same message. This CL
fixes the problem by not creating a thread for notification_ind messages
when autodownload is enabled. Messages will show up when they're completely
downloaded.

Change-Id: If515f134026bafaa2256673de1cff89976df8f59