ANR in messaging after launch and clicking on an existing thread
Bug 6772676

The UI thread was calling PhoneNumberUtil.isPossibleNumber which ends up
calling PhoneNumberUtil.getMetadataForRegion which is blocked on a sync
while a data file is loaded. In this case, our loading the contacts in a non-UI
thread associated with all the conversations happened to get there first.
Loading the file was slow enough to cause the blocked UI thread to ANR.
We can never call a file loading function on the UI thread. Since this
call is simply coming from context menu code, deciding whether to show an Add contact
menu item for the number, I'm now using Telephony.Mms.isPhoneNumber which doesn't result in a file getting
loaded on the UI thread.

Change-Id: I4bc1db22c6e20fd78e988fa7e56b7cb37a5cc62e