Notification goes to new message instead of existing thread
Bug 7169881

blockingUpdateAllNotifications() gets called from Conversation.markAsRead().
markAsRead gets called often, on every DB change (see CMP.mDataSetChangedListener)
and on every user interaction (see CMP.onUserInteraction()).
The call to blockingUpdateAllNotifications() wasn't passing in the threadId.
As a result, the code would build a notification for an incoming message
even though CMP was looking at that thread. This caused the messaging
notification icon to flash in notification bar as notification were
created then quickly dismissed. This change fixes that by not creating
notifications (other than a soft beep) for incoming messaging on the
thread the user is looking at. It also fixes the empty notification that
would take the user to a new conversation.

Change-Id: Ic0177cdd05d198abc85eae9077688c360573c879