Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

Thanks for noticing this problem and going through all the work. It turns out there was an internal issue that caused those icons not to be generated. That process has been fixed and all the icons (of the various resolutions) will be in the next release.

Patch-set: 2