Fix notification icon flashing
While testing messaging notifications, I noticed there's a lot of flashing
of the mms icon in the notification bar. I didn't even have to look at the
code to know why. blockingUpdateNewMessageIndicator() gets called from a
number of places to update the notifications. Before rebuilding notifications
it first cancels any existing notifications. This causes the flash. A
simple solution that eliminates the flash is to not cancel the notification
if we're going to put up a new notification.

Change-Id: I7a12508d9e7c6725bbd581e231aa6dda090ec971