Add setting to turn on/off group MMS
Bug 6970173

Add a setting/preference to turn on/off group MMS. If group mms is disabled
by mms_config.xml, the preference is hidden from the user (this is consistent
with other mms features disabled by carriers). When group mms is disabled,
receiving a group mms is also disabled. In other words, the received message
will appear to have come from a thread with only the sender and not in a thread
with the sender + the other recipients.

Change-Id: I9510ae269f7b3968c768c7b5f5cf8dcc7fe4eca6