b/7363553: part 2 of fix keyboard and window layout interaction when re-entering ComposeMessageActivity.
(this is on top of the cl https://googleplex-android-review.googlesource.com/#/c/242627/4)

The simple idea is to delay loading the message history and draft text until the keyboard
is shown. The effect is, the compose screen shows blank, keyboard comes up, and the message
history and draft get displayed.

While the idea is simple, a lot has to be changed in ComposeScreenActivity to make this happen.

- message history query and draft loading were done separately before. Changed to perform
those actions together. If the keyboard needs to be shown, delay the loading of message+draft
until we detect the keyboard is shown the first time. The detection is done in the windowMgr
callback of onSizeChanged(). If we detect the new window height is less than the old height by
200 pixels, we treat it as "keyboard opened event".

the decision to show keyboard (plus delay message+draft loading) is done in onStart(), but
we set the window mode in onResume(), and hide the keyboard in onPause(). Hiding the keyboard
in onPause() makes it less janky to show keyboard when re-entering the activity.

- modify the initialization of the compose view to not display draft text (which can be
in the compose textfield w/o loading a draft), until the keyboard comes up and we are ready
to load the message+draft.

- remove conditional draft loading code in onRestart(). We always load the draft in onStart() now.
loadDraft() can be a no-op if we come back through onRestart(), if the draft is in memory
already.

- call loadMessageAndDraft in onNewIntent(), which doesn't go thru. onStart() code path.

Change-Id: I2ae953815fe2325b7038a8c647a415275cf77b5f