Empty draft created on deleting a group MMS message
Bug 7360609

Don't save an mms draft for a group message when there's no text. When the
last message in a thread is deleted, automatically go back to the
ConversationList. Earlier when my deletes were failing because of the
NULL threadId in the pdu table, I noticed an exception thrown from
Conversation.startDelete: "startDeleteAll already in the middle of a delete".
startDelete was being called in a loop and of course sDeletingThreads would
be true if the delete hadn't completed in between iterations. Fix by having
startDelete do all the deleting of multiple threadIds.


Change-Id: Idfb8c0d54d8bb1875bbe7c5163e75bbb0dfbf14d