Fix to hide keyboard jank-free when coming back to MMS with no draft.
What was happening is that if you loaded a conversation, brought
up the IME, left, then came back, there was a brief period where the
window size was still small (leaving room for the IME), then the IME
would go away and the window size would expand.

Hiding the keyboard in onPause or onStop didn't prevent this. So
the fix is to extend the hack of delaying loading the draft messages
to all cases, not just the case where the keyboard is going to be
shown. This way the window manager settles down before we load
the content, and everything comes into view in the right place the
first time.

b/7481258

Change-Id: If2592737bda4d2938fc8fc982724459d3d26fd35