Opening recent apps while on compose dismisses compose on re-rentering app
Bug 7586014

Don't bail back to ConversationList when we've got an empty ComposeMessageActivity.
In the case where the user has entered a recipient and followed the same
repro steps in the bug, a draft is created (see CMA's saveDraft) because
the receipient editor is open and there is a recipient. In that case, the
working message is not discarded and does not go through the escape-to-
ConversationList case. By the way, I was not able to repro the original
bug 2793874 that added this escape clause code.

Change-Id: I92d84f19c8ba2c27a8c28975dc95d1d230b89f1c