Update patch set 1
Patch Set 1:

This is the fix for the issue mentioned in https://android-review.googlesource.com/44413

COPY REPRO STEPS FROM Change#44413
Repro Steps: 1. Create a conversation which has a Mms to A 2. Make sure there aren't any drafts in the conversation 3. Long press the Mms in the conversation to forward 4. Input A's number to forward the Mms to A Expect: After 4, finish the forward activity and turn back to original conversation(A's) without any drafts. Actual: After 4, Back to A's conversation and kept the forwarded Mms as the draft.

Patch-set: 1
Reviewer: Roger Chen <1005243@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0