The message details performs wrong type
When user want to check the detailed information of a mms.
The messages are divided into two classes: mms and sms. One
mms and one sms maybe have the same id but have different
message type. Add a type check in getCursorForItem function,
so we can get the right cursor.

Change-Id: I48cbc9e437b1c33788efceff49acc679151dcb63
Signed-off-by: Rui Wang <ruix.wang@intel.com>
Signed-off-by: Bin Yang <bin.y.yang@intel.com>