Use the correct host route function in MMS.
Currently, MMS sets up host routes using requestRouteToHost().
This function is deprecated, and takes an int, so MMS has a
lookupHost() function that converts an IP addresses to an int.
lookupHost has a bug that returns corrupt data in the presence
of IPv6 addresses, so MMS doesn't work over IPv6.

Fix this by using the new requestRouteToHostAddress instead of
requestRouteToHost, and delete the function.

This shouldn't break anything, because the result of the lookup
is never used outside the function, it's only passed to
ConnectivityService to set up the host route, so the rest of the
MMS code won't see any change in behaviour.

Bug: 8276725
Change-Id: If6656a33fdea01c6fba178b34ef818d48cf7a395