Video capture is returning a much larger result than specified in android.intent.extra.durationLimit
Bug 8540985

We need to truncate the scrap file before writing to it, otherwise the
camera will end up writing to the previous file, leaving the previous
length if shorter. If you captured a pictures (1.2m) and then replaced
the picture with a short video (a couple hundred k), the video file
would still be 1.2m.

Change-Id: I56c913d7ac530ec5a8c5f330bc7f9aa08512c48e