Unable to playback the captured video in Messaging app, displays "Can't play this video"
Bug 9193742

The TempFileProvider's getTempStoreFd needs to respect the file mode.
Before this change, the file was always truncated when opened. When the
user tried to playback the freshly captured video in the video camera, the
video player would go through the provider to open the file to playback.
However, the file was truncated so nothing could be played back. With this
change, the file isn't truncated when opened for just reading.

Change-Id: Ifcfe8d586c345271186856c61d5f6e4fa32d130e