Not all MMS is sent or received.
When there are multiple sending and receiving transactions,
there is possibility if a case in which there are two
transactions in processinglist and only one will be
processed.

This patch adds the transaction to the pending list when
there is already a transaction in processing state and also
doesn't disconnect mms connectivity if there are
transactions in processing state.

Change-Id: I77c60726dbdb43280057bd3db60d9649b15a17f2
Signed-off-by: Rui Wang <ruix.wang@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>