Add contact sharing through SMS
This extension shares contact information through SMS message, using a text data
encapsulation protocol, and facilitates receiving and saving this kind of
information by opening the New Contact activity with its fields already
completed. Also, when sharing a contact the user can choose which information
to share through the application.
Contact sharing was integrated in the Mms by inserting classes:Encapsulation
in mms.model, ArrayOptions, ListItemComponents, OptionsList and Dialog in
mms.ui, CheckIsContact and ReceiveContend in mms.transaction and by merging
sending the contact information in ComposeMessageActivity and receiving the
information in PrivilegedSmsReceiver.

Change-Id: I2dcb52bf184091dbbec9c1408f12adc38bcb518e
Signed-off-by: Maria Constantin <maria.p.constantin@intel.com>
Signed-off-by: Alina Rublea <alina-mirela.rublea@intel.com>
Co-authored-by: Alina Rublea <alina-mirela.rublea@intel.com>