MMS: Create an entry in thread table for WAP PUSH
Messaging application displays the SMS/MMS based on
thread id. If WAP PUSH is present in DB but the actual
retrieval of MMS failed then Messaging app would not
show that message and hence user would never get a chance
to manually download that MMS. Hence is it required that
for every WAP PUSH we create thread id.

Change-Id: I7e4b16072a8f2cd5d1424cd8b476ddd98cb9fbeb