The messaging app needs to look at getMobileDataEnabled
10456183 - Messaging app is deemed responsible for 20-30% of the battery drain

If the user has turned off Mobile network settings->Data enabled, then
it's impossible to download an mms. However, the TransactionService would
keep on trying to get the connection manager to switch to the mms mobile
state, hence the strong grip on the wakelock. With this change, consider
the phone as not connected if data is disabled.

Change-Id: Ie4b359bb3e05472f187d79bf42dbb43f21eb31f3