Mms: Do not call stopUsingNetworkFeature when another MMS is in process
Bug 10625264

Around the time when a big MMS transaction (send or received) is about
to complete, a new MMS transaction (send or received) is started.
Due to completion of the previous transation, present code logic
removes the route by calling stopUsingNetworkFeature(). This
results in socket failure for the new MMS transaction. From contribution:
https://android-review.googlesource.com/#/c/63809/2

Change-Id: I501edf522ba957d3ff39e77a1cd0e73c49d2d595