Update Mms app to better handle not being the default app
Adds new settings to make it easier to make this the default
app and to make it easy to change the default app.

Disables all options and menus in the UI which would send
or otherwise perform write operations on the SMS database when the app is
not the default SMS application.

Enables or disables all settings UX based on whether or
not it is the default SMS application.

Adds a banner to the top of the conversation list if Messaging is not
the default SMS app making it easy to set Messaging as the default app.

Bug: 10819173 change Messaging app to handle the use case when it's not the default SMS app on KLP
Change-Id: I68dc1c0c72df75bda23f3a137ccc3b279526bb2c