Mms app should stop sending messages if it is not the default app.
If the Mms app had pending (Sms or Mms) messages when the user changed the
default Sms app to another app, it would try to send the pending messages
over and over because it would send and then fail to mark the message as
sent, send, fail to mark, etc.

This change makes the SMS receiver servce (which actually sends SMS as well
as receiving them) bail if the Mms app is not the default app and does the
same for the TransactionService (Mms). The message already in flight will
still get sent once, but the service will not try again or send any other
messages.

If the user then makes the Mms app the default app again these messages
will be sent once more (whether or not they were sent the first time,
because we can't tell) and then they will be marked as sent. The user may
need to send an SMS/MMS message again to "kick the queue" and cause the
pending messages to be sent after disabing/reenabling the Mms app as the
default SMS app.

Bug: 10955896 Phone sends an infinite amount of MMS messages if you switch default SMS app before MMS sends
Change-Id: I35c895cdf9ef7c597d1a25d7c3b8a801e0226688