Update Messaging app with msim changes from klp-sprout
Update the Messaging app to use the new klp-sprout msim APIs. The app
can serve as a test-bed for msim while we bring up the new sms app.
Also, it will serve as an example for others using msim. The code for
making a msim connection for mms is commented out for now
(still uses the old api) as that api hasn't been implemented for
l-sprout-dev yet.

Change-Id: I698cd5c2959aa7e270294386a9bcf77c707b201f