Messaging app needs to use mms_config appropriate for each SIM
Bug 18018147

The sprout Messaging app only loads a single mms_config, even when there
are two sims. We didn't fix this in klp-sprout because we didn't have
specific mms_configs for the mcc/mnc's we were launching in. Hence,
the app would use the default mms_config values.

In L we made the mms service load individual mms_config's for each
sim and provided a way to get the config values based on the subscription
id. Since we're making the old Messaging app available in L-mr1 for AOSP,
we need to fix the app to use config values based on the subId. In
addition, we need to fix all the references to getLine1Info to get the
appropriate number based on the subscription id.

Change-Id: Ib7ffab8faa24f307e9f4877ab5292e23f9587dd3